Move the script requirements after description of stock tasmota firmware
I think what was being implied is that the Tasmota FW (with specified template) will control the TUYA MCU via standard serial.  So that one can MANAGE the switch and dimming via the tasmota interface.  (btw this proved to be incorrect for me and the Gosund SW2 i bought on 27AUG20 was able to flash, and I could control on/off; but was not able to control dimmer).

But the reason for this proposed change is because i think the original author wants to imply that the device is "controllable" from Tasmota interface, but that the devices' capacitive dimmer will NOT work unless "scripting" is enabled.  So scripting should come after the base description.

Also, there is a link to the "scripting" but it seemed to land on standard tasmota splash page as opposed to talking about "how to enable scripting" in the tasmota code.

This is my first contribution, and I hope it is correct (i.e. that there is not somthing I misunderstood), and I hope it adds value to the docs.  If not I am sorry for wasting your time.